requirements.txt
fastapi
uvicorn
sqlalchemy==2.0.15
psycopg2
python-dotenv
python-multipart
pydub